pyproject.toml
[project]
name = "gotale-backend"
version = "0.1.2"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "astroid==3.3.9",
    "asttokens==3.0.0",
    "baker==1.3",
    "certifi==2025.1.31",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "decorator==5.2.0",
    "dill==0.3.9",
    "distlib==0.3.9",
    "django==5.1.6",
    "django-extensions==3.2.3",
    "django-filter==25.1",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.4.0",
    "djangorestframework-stubs==3.15.3",
    "drf-rw-serializers>=1.4.0",
    "drf-spectacular>=0.28.0",
    "executing==2.2.0",
    "factory-boy==3.3.3",
    "faker==36.1.1",
    "filelock==3.18.0",
    "git-filter-repo==2.47.0",
    "identify==2.6.9",
    "idna==3.10",
    "inflection==0.5.1",
    "iniconfig==2.0.0",
    "jedi==0.19.2",
    "markdown==3.7",
    "markdown-it-py==3.0.0",
    "matplotlib-inline==0.1.7",
    "mccabe==0.7.0",
    "mdurl==0.1.2",
    "model-bakery==1.20.4",
    "nodeenv==1.9.1",
    "packaging==24.2",
    "parso==0.8.4",
    "pathspec==0.12.1",
    "pexpect==4.9.0",
    "platformdirs==4.3.7",
    "pluggy==1.5.0",
    "pprintpp==0.4.0",
    "prompt-toolkit==3.0.50",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rest-framework-simplejwt==0.0.2",
    "rich==14.0.0",
    "sqlparse==0.5.3",
    "stack-data==0.6.3",
    "tomlkit==0.13.2",
    "traitlets==5.14.3",
    "types-pyyaml==6.0.12.20250402",
    "types-requests==2.32.0.20250328",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "virtualenv==20.30.0",
    "wcwidth==0.2.13",
]


[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "W",  # pycodestyle warnings
  "UP", # pyupgrade (modernize code)
  "I",  # isort (import sorting)
]
ignore = ["E501"] # Example: ignore line-length errors

exclude = [
  "migrations/",  # Django migrations
  ".venv/",
  "__pycache__/",
]

[tool.ruff.lint.isort]
known-first-party = ["backend"] 

[tool.pylint.MASTER]
load-plugins = 'pylint_django'

[dependency-groups]
dev = [
    "django-stubs==5.1.3",
    "django-stubs-ext==5.1.3",
    "icdiff==2.0.7",
    "ipython==8.32.0",
    "isort==6.0.1",
    "pre-commit==4.2.0",
    "pytest>=8.3.4",
    "pytest-clarity==1.0.1",
    "pytest-django==4.10.0",
    "pytest-factoryboy==2.7.0",
    "pytest-icdiff==0.9",
    "pytest-mock==3.14.0",
    "pytest-unordered>=0.7.0",
    "ruff==0.13.0",
    "werkzeug>=3.1.3",
]
